Fixed some error
I found out that I was working in the wrong component, I was working in the Product.js instead of Products.js. I created the CartContext.js component in the contexts folder. I added and imported inside of that component and then I took out the props in the ShoppingCart.js and Navigation.js. But not got an error about cart being undefined.